refactor: movendo ClientProfileHeader para a feature clients
